# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Loftly API — Phase 1 MVP
  description: |
    The loftly-api HTTP contract for Phase 1. See `/mvp/API_CONTRACT.md` for the
    reader-friendly guide; this file is the source of truth for code generation
    and runtime validation.

    Base path: `/v1`. All timestamps are RFC 3339 with timezone. Monetary values
    are in THB as `numeric` strings to avoid float precision loss.
  version: 0.1.0
  contact:
    name: Loftly
    url: https://loftly.co.th
servers:
  - url: https://api.loftly.co.th/v1
    description: Production
  - url: https://staging-api.loftly.co.th/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Local dev
...
security:
  - {}  # default: no auth; per-endpoint overrides below
paths:
...
  /selector:
    post:
      tags: [selector]
      summary: Submit spend profile; receive ranked card stack
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SelectorInput' }
      responses:
        '200':
          description: Result (initial JSON envelope; additional SSE frames stream rationale chunks)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelectorResult' }
            text/event-stream:
              schema:
                type: string
                description: SSE stream of rationale chunks then final `done` event
        '422':
          description: Invalid profile (category sum mismatch, goal mismatch, etc.)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Rate-limited
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: All LLM paths degraded (rule-based fallback served)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelectorResult' }
  /selector/{sessionId}:
    get:
      tags: [selector]
      summary: Retrieve previously computed result
      security: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: token
          in: query
          required: true
          schema: { type: string }
          description: Signed token (from magic link or initial response)
      responses:
        '200':
          description: Cached result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelectorResult' }
...
  /promos:
    get:
      tags: [promos]
      summary: Active promotions
      security: []
      parameters:
        - name: bank
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: card_id
          in: query
          schema: { type: string, format: uuid }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Promo list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PromoList' }
...
  /admin/promos:
    get:
      tags: [admin]
      summary: List promos
      security: [{ adminJWT: [] }]
      responses:
        '200':
          description: Promo list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PromoList' }
    post:
      tags: [admin]
      summary: Create manual promo
      security: [{ adminJWT: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PromoUpsert' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Promo' }
...
components:
  securitySchemes:
    userJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT with role=admin claim
    internalApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    webhookHMAC:
      type: apiKey
      in: header
      name: X-Loftly-Signature
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message_en]
          properties:
            code: { type: string }
            message_en: { type: string }
            message_th: { type: string }
            details: { type: object, additionalProperties: true }
    Pagination:
      type: object
      properties:
        cursor_next: { type: string, nullable: true }
        has_more: { type: boolean }
        total_estimate: { type: integer }
...
    Promo:
      type: object
      properties:
        id: { type: string, format: uuid }
        bank_slug: { type: string }
        external_source_id: { type: string, nullable: true }
        source_url: { type: string, format: uri }
        promo_type: { type: string, enum: [category_bonus, cashback, transfer_bonus, signup, statement_credit, dining_program] }
        title_th: { type: string }
        title_en: { type: string, nullable: true }
        description_th: { type: string, nullable: true }
        merchant_name: { type: string, nullable: true }
        category: { type: string, nullable: true }
        discount_type: { type: string, nullable: true }
        discount_value: { type: string, nullable: true }
        discount_amount: { type: number, nullable: true }
        discount_unit: { type: string, nullable: true }
        minimum_spend: { type: number, nullable: true }
        valid_from: { type: string, format: date, nullable: true }
        valid_until: { type: string, format: date, nullable: true }
        card_ids:
          type: array
          items: { type: string, format: uuid }
        active: { type: boolean }
    PromoList:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Promo' }
        pagination: { $ref: '#/components/schemas/Pagination' }
    PromoUpsert:
      type: object
      additionalProperties: true
...
    SelectorInput:
      type: object
      required: [monthly_spend_thb, spend_categories, goal, locale]
      properties:
        monthly_spend_thb:
          type: integer
          minimum: 5000
          maximum: 2000000
        spend_categories:
          type: object
          description: keys in {dining, online, travel, grocery, petrol, other}
          additionalProperties: { type: integer }
        current_cards:
          type: array
          items: { type: string, description: cards.slug }
        goal:
          type: object
          required: [type]
          properties:
            type: { type: string, enum: [miles, cashback, benefits] }
            currency_preference: { type: string, nullable: true }
            horizon_months: { type: integer, nullable: true }
            target_points: { type: integer, nullable: true }
        locale: { type: string, enum: [th, en] }
    SelectorResult:
      type: object
      properties:
        session_id: { type: string, format: uuid }
        stack:
          type: array
          items:
            type: object
            properties:
              card_id: { type: string, format: uuid }
              slug: { type: string }
              role: { type: string, enum: [primary, secondary, tertiary] }
              monthly_earning_points: { type: integer }
              monthly_earning_thb_equivalent: { type: integer }
              annual_fee_thb: { type: number, nullable: true }
              reason_th: { type: string }
              reason_en: { type: string, nullable: true }
        total_monthly_earning_points: { type: integer }
        total_monthly_earning_thb_equivalent: { type: integer }
        months_to_goal: { type: integer, nullable: true }
        with_signup_bonus_months: { type: integer, nullable: true }
        valuation_confidence: { type: number }
        rationale_th: { type: string }
        rationale_en: { type: string, nullable: true }
        warnings:
          type: array
          items: { type: string }
        llm_model: { type: string }
        fallback: { type: boolean }
        partial_unlock:
          type: boolean
          description: true when anonymous user sees only primary
...